requirements.txt

fastapi>=0.115.0
uvicorn[standard]>=0.32.0
jinja2
python-multipart
sqlalchemy
databases[sqlite]
pydantic>=2.0.0
livekit>=0.11.0
ruff
python-dotenv
alembic
